% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bias}
\alias{bias}
\title{Calculate model bias}
\usage{
bias(obs, pred)
}
\arguments{
\item{Calculates}{model bias, given 2 vectors of observed and predicted values}
}
\description{
Calculate model bias
}
\examples{
bias()
}
\keyword{bias,}
\keyword{fit}
\keyword{goodness}
\keyword{of}

