% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cache.check}
\alias{cache.check}
\title{Check that the cache directory location has been set}
\usage{
cache.check()
}
\description{
Checks that the cache directory location has been set.  Also checks that the directory contains all necessary subdirectories
}

