% Generated by roxygen2 (4.0.1): do not edit by hand
\name{capitalize}
\alias{capitalize}
\title{Customization of capitalize function}
\usage{
capitalize(string)
}
\arguments{
\item{Simple}{function to capitalize first letter or string, for prettying up plot titles, legends, axes, etc.}
}
\description{
Customization of capitalize function
}
\keyword{capitalize}
\keyword{string,}

