% Generated by roxygen2 (4.0.1): do not edit by hand
\name{char.columns.retrieve}
\alias{char.columns.retrieve}
\title{x}
\usage{
char.columns.retrieve(id = F, impoundment = F)
}
\arguments{
\item{x}{\code{x}}

\item{x}{\code{x}}

\item{x}{\code{x}}
}
\value{
x \code{x}
}
\description{
x
}
\seealso{
x \code{\link{conteStreamflow::}}, \code{\link{conteStreamflow::}}
}

