% Generated by roxygen2 (4.0.1): do not edit by hand
\name{char.retrieve}
\alias{char.retrieve}
\title{Get basin characteristics associated with each gage}
\usage{
char.retrieve(gages.spatial, cols = NULL, log.cols = NULL,
  std.cols = NULL, log.std.cols = NULL, return.scaling = F, impound = F,
  server.url = "http://felek.cns.umass.edu:9283")
}
\arguments{
\item{gages.spatial}{\code{SpatialPointsDataFrame} of gage info from plot.gages.nhdplus()}

\item{basin.char.file}{\code{character} location andname of .rdata file that contains UpstreamStats, by NHDplus FEATUREID}
}
\value{
\code{SpatialPointsDataFrame} of gage info including upstream basin characteristics
}
\description{
Get basin characteristics associated with each gage, or more precisely, with the NHDplus stream reach each gage is plotted to.
}
\seealso{
plot.gages.nhdplus, load.gage.impound
}

