% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.q.matrices}
\alias{create.q.matrices}
\title{create the monster list of 3-d matrices for flow}
\usage{
create.q.matrices(gages.spatial, periods = c("daily", "monthly", "seasonal",
  "annual"), template.date = NULL)
}
\description{
used to store aggregated flow data
}

