% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.template.date.mauer}
\alias{create.template.date.mauer}
\title{Create template of dates in mauer weather data}
\usage{

  create.template.date.mauer(create.template.periods = (conteStreamflow::create.template.periods))
}
\description{
mauer dates
}

