% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cv.error}
\alias{cv.error}
\title{Calculate coefficient of variation of error}
\usage{
cv.error(obs, pred)
}
\arguments{
\item{Calculates}{CV of error, given 2 vectors of observed and predicted values}
}
\description{
Calculate coefficient of variation of error
}
\examples{
cv.error()
}
\keyword{cv,}
\keyword{fit}
\keyword{goodness}
\keyword{of}

