% Generated by roxygen2 (4.0.1): do not edit by hand
\name{filter.gage.impound}
\alias{filter.gage.impound}
\title{Filter out gages based on upstream barriers and impoundments}
\usage{
\method{filter}{gage.impound}(gages.spatial)
}
\arguments{
\item{gages.spatial}{\code{SpatialPointsDataFrame} of gage info including column names "TNC_DamCount" and "OnChannelWaterSqKM"}
}
\value{
\code{SpatialPointsDataFrame} of gage info that meet default barrier and impoundment criteria
}
\description{
Using barrier and impoundment info returned by load.gage.impound()
}
\seealso{
load.gage.impound
}

