% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flow.retrieve}
\alias{flow.retrieve}
\title{import flow}
\usage{
flow.retrieve(gages.spatial, periods = c("seasonal", "annual"),
  agg.function.flow = (conteStreamflow::agg.function.flow),
  template.date = NULL, template.period = NULL, cols.flow = NULL)
}
\description{
import flow from nwis web service, and aggregate to various metrics and by various periods
}

