% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.buffer}
\alias{gage.buffer}
\title{Internal function to filter gages based on whether they fall within a geographic buffer}
\usage{
gage.buffer(gages.spatial, plot = F, buffer = NULL, buffer.file = NULL,
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs",
  message = FALSE)
}
\arguments{
\item{gages.spatial}{\code{SpatialPointsDataFrame}}

\item{buffer}{\code{SpatialPolygonsDataFrame} must provide either a SpatialPolygonsDataFrame buffer, or a file name and location of a shapefile of a single polygon to outline area of interest}

\item{buffer.file}{\code{character}}

\item{proj4}{character  coordinate system, in proj4 syntax.  Defaults to GCS NAD83, which is used by NHDplus}

\item{plot}{\code{boolean}  if true, will plot a map of gages with state outlines}

\item{message}{\code{boolean} if true, will display a message indicating how many gages were removed}
}
\value{
SpatialPointsDataFrame
}
\description{
This function is used internally by gage.retrieve and doesn't need to be called explicitly by the user for normal uses.  Input spatial gage object is overlaid with buffer polygon, and gages falling outside the buffer are filtered out.
}
\seealso{
\code{\link{conteStreamflow::gage.retrieve}}
}
\keyword{SpatialPointsDataFrame}
\keyword{gage,}

