% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.get}
\alias{gage.get}
\title{Get gage location and information from NWIS}
\usage{
gage.get(states = c("CT", "ME", "MA", "NH", "NY", "RI", "VT"),
  max.da.sqkm = 50, min.da.sqkm = 0,
  temp.dir = "C:/ALR/Models_processed_data/flow_timeseries", save.log = T,
  log.dir = "C:/ALR/Models_processed_data/flow_timeseries")
}
\arguments{
\item{states}{\code{character vector} of state abbreviations}

\item{max.da.sqkm}{\code{numeric} filter gages by min and max drainage area}

\item{min.da.sqkm}{}

\item{temp.dir}{\code{character} directory to save temporary files and gage retrieval logs/metadata}

\item{save.log}{\code{boolean} whether to save a log}
}
\value{
\code{data.frame}
}
\description{
Get gage information, including site_no and lat/long coordinates from NWIS, by specifying a list of states to search within.
}
\keyword{gage}
\keyword{nwis,}

