% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.load.char}
\alias{gage.load.char}
\title{Get basin characteristics associated with each gage}
\usage{
gage.load.char(gages.spatial,
  basin.char.file = "C:/ALR/Models_from_others/Kyle/NH NY NHDplus basin zonal stats KO/NENY_CovariateData_2014-01-23.RData")
}
\arguments{
\item{gages.spatial}{\code{SpatialPointsDataFrame} of gage info from plot.gages.nhdplus()}

\item{basin.char.file}{\code{character} location andname of .Rdata file that contains UpstreamStats, by NHDplus FEATUREID}
}
\value{
\code{SpatialPointsDataFrame} of gage info including upstream basin characteristics
}
\description{
Get basin characteristics associated with each gage, or more precisely, with the NHDplus stream reach each gage is plotted to.
}
\seealso{
plot.gages.nhdplus, load.gage.impound
}

