% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.place.nhdplus}
\alias{gage.place.nhdplus}
\title{Assigns each gage to a NHDplus catchment/stream reach}
\usage{
gage.place.nhdplus(gages.spatial)
}
\arguments{
\item{gages.spatial}{\code{SpatialPointsDataFrame} usually produced by gage.retrieve()}

\item{cache.dir}{\code{character} local directory where spatial data files can be cached}
}
\value{
\code{SpatialPointsDataFrame}
}
\description{
Uses spatial gage object, usually one created by gage.retrieve().
Loads NHDplus catchment file, determines which catchment each gages falls within, and assigns a FEATUREID
}
\seealso{
\code{\link{conteStreamflow::gages.retrieve}}
}
\keyword{NHDplus}
\keyword{gage,}

