% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.place}
\alias{gage.place}
\title{Internal function to create SpatialPointsDataFrame from data.frame of gage information}
\usage{
gage.place(gages.df, lat_column = "dec_lat_va", long_column = "dec_long_va",
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs", plot = F)
}
\arguments{
\item{gages.df}{\code{data.frame} of gage information, including columns with lat and long coordinates}

\item{lat_column}{\code{character} defaults to NWIS column name "dec_lat_va", but can be customized for use with a different data source}

\item{lat_column}{\code{character} defaults to NWIS column name "dec_lat_va", but can be customized for use with a different data source}

\item{proj4}{\code{character}  coordinate system, in proj4 syntax.  Defaults to GCS NAD83, which is used by NHDplus}

\item{plot}{\code{boolean}  if true, will plot a map of gages with state outlines}
}
\value{
\code{SpatialPointsDataFrame}
}
\description{
This function is used internally by gage.retrieve and doesn't need to be called explicitly by the user for normal uses.  This function uses lat and long coordinates from gage info data frame, and turns it into a spatial object w/ the data attached.
}
\seealso{
\code{\link{conteStreamflow::gage.retrieve}}
}
\keyword{SpatialPointsDataFrame}
\keyword{gage,}

