% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.place.weather.grid}
\alias{gage.place.weather.grid}
\title{plot gages to weather grid}
\usage{
gage.place.weather.grid(gages.spatial, plot = F)
}
\description{
plot to a weather grid polygon, based on all the files in the mauer daily east data set
}

