% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.plot}
\alias{gage.plot}
\title{Create SpatialPointsDataFrame from data.frame of gage information}
\usage{
gage.plot(gages.df,
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs")
}
\arguments{
\item{gages.df}{\code{data.frame} of gage information, including columns titled "dec_long_va" and "dec_lat_va"}

\item{proj4}{\code{character}  coordinate system, in proj4 syntax.  Defaults to GCS NAD83, which is used by NHDplus}
}
\value{
\code{SpatialPointsDataFrame}
}
\description{
Uses gage info created by get.gages, uses the columns with lat and long coordinate info to turn it into a spatial object (SpatialPointsDataFrame), preserving all data in the data.frame
}
\seealso{
get.gages
}
\keyword{SpatialPointsDataFrame}
\keyword{gage,}

