% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.plot.weather}
\alias{gage.plot.weather}
\title{plot gages to a weather grid cell}
\usage{
gage.plot.weather(gages.spatial, weather.grid.poly = NULL,
  weather.grid.dir = "C:/ALR/Data/ClimateData/Mauer/met_poly_GCS",
  weather.dir = "C:/ALR/Data/ClimateData/Mauer/daily/east",
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs", plot = F)
}
\description{
plot to a weather grid polygon, based on all the files in the mauer daily east data set
}

