% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gage.trace.dams}
\alias{gage.trace.dams}
\title{iteratively trace nhdplus network upstream and cound # and size of dams}
\usage{
gage.trace.dams(gages.spatial,
  dam.file = "C:/ALR/Data/ConnectivityData/TNC_100k/dams_on_med_rez.dbf",
  plusflow.file = "C:/ALR/Data/StreamData/NHDplus/NHDPlusAttributes/0205/PlusFLow.dbf")
}
\description{
for gages that don't have dam counts and impounded area in basin characteristics only
}

