% Generated by roxygen2 (4.0.1): do not edit by hand
\name{import.flow}
\alias{import.flow}
\title{import flow}
\usage{
import.flow(gages.spatial, periods = c("seasonal", "annual"), save.log = T,
  log.dir = "C:/ALR/Models_processed_data/flow_timeseries",
  template.date = NULL)
}
\description{
import flow from nwis web service, and aggregate to various metrics and by various periods
}

