% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.gage.impound}
\alias{load.gage.impound}
\title{Get information on upstream barriers and impoundments}
\usage{
load.gage.impound(gages.spatial,
  basin.impound.file = "C:/ALR/Models_from_others/Kyle/NH NY NHDplus basin zonal stats KO/NENY_CovariateData_Impoundments_2014-01-23.RData")
}
\arguments{
\item{gages.spatial}{\code{SpatialPointsDataFrame} of gage info from plot.gages.nhdplus()}

\item{basin.impound.file}{\code{character} location and name .Rdata file that contains UpstreamStatsImpounded, by NHDplus FEATUREID}
}
\value{
\code{SpatialPointsDataFrame} of gage info including TNC barriers and area of wetlands and open water
}
\description{
Load and attach information for each gage on upstream barriers and impoundments.  Uses TNC barrier inventory and FWS survey openwater and wetlands.
(Open water and wetlands are intersected with NHDplus med res flowlines, to differentiate on- and off-network wetlands and waters)
}
\seealso{
plot.gages.nhdplus, load.gage.char, filter.gage.impound
}

