% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.sp}
\alias{merge.sp}
\title{internal, temporary replication of merge function to explicitly indicate merging of spatial object with a data.frame}
\usage{
merge.sp(x, y, by = intersect(names(x), names(y)), by.x = by, by.y = by,
  all.x = TRUE, suffixes = c(".old", ""), incomparables = NULL, ...)
}
\description{
perhaps remove?
}

