% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.gages.buffer}
\alias{plot.gages.buffer}
\title{Filter gages based on whether they fall within a geographic buffer}
\usage{
plot.gages.buffer(gages.spatial, plot = F,
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs",
  buffer.file = "C:/ALR/Model_processed_dataflow_gages/ctr_states_buffer_no_islands2")
}
\arguments{
\item{gages.spatial}{SpatialPointsDataFrame produced by plot.gages() function}

\item{proj4}{character  coordinate system, in proj4 syntax.  Defaults to GCS NAD83, which is used by NHDplus}
}
\value{
SpatialPointsDataFrame
}
\description{
Uses spatial gage object (SpatialPointsDataFrame) created by plot.gages(),
and filters them based on whether they fall within a buffer.  Default buffer around CT River basin states is specified.
}
\seealso{
get.gages
}
\keyword{SpatialPointsDataFrame}
\keyword{gage,}

