% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save.log}
\alias{save.log}
\title{internal function to save log file}
\usage{
save.log(text, filename, ext = "txt")
}
\arguments{
\item{text}{\code{character} text for writeLines}

\item{dir}{\code{character} name of directory location}

\item{filename}{\code{character} file name prefix}
}
\value{
ext \code{character} file extension
}
\description{
saves log file in directory specified.  uses specified prefix.  reads names of existing logs in the directory to specify a number, incrementing it from the largest number of that file type.
}

