% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.na}
\alias{summary.na}
\title{Customization of summary function}
\usage{
summary.na(v, hist = F)
}
\arguments{
\item{\code{numeric}{vector}}
}
\value{
only prints summary
}
\description{
Customization of summary function that includes a summary of the number of values that are na, infinite, or not numeric.  Also reports, n records, min, mean, max, and several quantiles.
}
\keyword{na}
\keyword{summary,}

