% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to.month}
\alias{to.month}
\title{Associates calendar dates with a date representing its month.}
\usage{
to.month(d)
}
\arguments{
\item{d}{\code{vector of dates}}
}
\value{
\code{vector of dates}
}
\description{
Given a date, determines which month it falls into, and assigns the date corresponding to the first calendar day of the month.  Yeah, this one is probably not necessary, but it's easier to have a function analogous to season and water.year assignment functions
}
\seealso{
\code{\link{to.water.year}}, \code{\link{to.season}}
}

