% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to.season}
\alias{to.season}
\title{Associate calendar dates either a date or string representing its season.}
\usage{
to.season(d, return.type = "date", input.type = "date")
}
\arguments{
\item{d}{\code{date} vector of dates}

\item{return.type}{\code{character} one of "date", "season", or "all".
"date" will return a vector of date object (which includes year)
"season" will return vector of strings of the season name (either "winter", "spring", "summer", or "fall") (does not include year)
"all" will return a data.frame, with date, string of season name, and numeric of year}
}
\value{
see above
}
\description{
Given a date, determines which season it falls into, and assigns it either a string of the season name,
}
\seealso{
\code{\link{conteStreamflow::to.water.year}}, \code{\link{conteStreamflow::to.month}}
}
\keyword{dates,}
\keyword{season}

