% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weather.grid.create}
\alias{weather.grid.create}
\title{create weather grid}
\usage{
weather.grid.create(weather.dir = "C:/ALR/Data/ClimateData/Mauer/daily",
  regions = c("east"),
  proj4 = "+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs",
  shapefile.dir = NULL)
}
\description{
create weather grid
}

