% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weather.retrieve}
\alias{weather.retrieve}
\title{import weather}
\usage{
weather.retrieve(gages.spatial, periods = c("seasonal", "annual"),
  create.cols.function = create.cols.weather,
  agg.function = agg.function.weather,
  weather.dir = "C:/ALR/Data/ClimateData/Mauer/daily", log.dir = NULL)
}
\description{
import weather
}

