% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bias}
\alias{bias}
\title{Calculate model bias}
\usage{
bias(obs, pred)
}
\arguments{
\item{obs}{\code{numeric vector} observed values}

\item{pred}{\code{numeric vector} predicted values (vector of same length as observed values)}
}
\value{
\code{list}
}
\description{
Calculates model bias, given 2 vectors of observed and predicted values
}
\seealso{
goodness
}
\keyword{bias,}
\keyword{fit}
\keyword{goodness}
\keyword{of}

