% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.w.matrices}
\alias{create.w.matrices}
\title{Create monster list of 3d weather matrices}
\usage{
create.w.matrices(weather.filenames, periods = c("daily", "monthly",
  "seasonal", "annual"), template.date, template.period, cols.weather)
}
\description{
used to store aggregated weather metrics
}

