% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cv.error}
\alias{cv.error}
\title{Calculate coefficient of variation of error}
\usage{
cv.error(obs, pred)
}
\arguments{
\item{obs}{\code{numeric vector} observed values}

\item{pred}{\code{numeric vector} predicted values (vector of same length as observed values)}
}
\value{
\code{numeric}
}
\description{
Calculates CV of error, given 2 vectors of observed and predicted values
}
\seealso{
goodness
}
\keyword{cv,}
\keyword{fit}
\keyword{goodness}
\keyword{of}

