% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flow.retrieve}
\alias{flow.retrieve}
\title{import flow}
\usage{
flow.retrieve(gages.spatial, periods = c("seasonal", "annual"),
  create.cols.function = create.cols.flow,
  agg.function = (headwaters::agg.function.flow), log.dir = NULL)
}
\description{
import flow from nwis web service, and aggregate to various metrics and by various periods
}

