% Generated by roxygen2 (4.0.1): do not edit by hand
\name{goodness}
\alias{goodness}
\title{Some goodness of fit, validation metrics}
\usage{
goodness(df = NULL, obs = NULL, pred = NULL)
}
\arguments{
\item{Calculates}{and prints a variety of goodness of fit, validation metrics, including rmse, NS eff, bias, percent bias, pearson's R, and cv of error}
}
\description{
Some goodness of fit, validation metrics
}
\examples{
goodness()
}
\keyword{fit}
\keyword{goodness}
\keyword{of}

