% Generated by roxygen2 (4.0.1): do not edit by hand
\name{goodness}
\alias{goodness}
\title{Some goodness of fit, validation metrics}
\usage{
goodness(obs.real = NULL, pred.log = NULL, bias.correction = NULL)
}
\arguments{
\item{obs}{\code{numeric vector} observed values}

\item{pred}{\code{numeric vector} predicted values (vector of same length as observed values)}
}
\value{
\code{data.frame}
}
\description{
Calculates and prints a variety of goodness of fit, validation metrics, including rmse, NS eff, bias, percent bias, pearson's R, and cv of error
}
\seealso{
bias, cv.error
}
\keyword{bias,}
\keyword{fit}
\keyword{goodness}
\keyword{of}

