% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ls.objects}
\alias{ls.objects}
\title{Customization of ls function}
\usage{
ls.objects(pos = 1, pattern, alpha = F, head = FALSE, n = 10)
}
\arguments{
\item{\code{}}{}

\item{\code{}}{}

\item{\code{}}{}
}
\value{
only print to console
}
\description{
Customization of ls function, that includes size (memory) of all objects or data in current session environment.
}
\examples{
ls.objects()
}
\keyword{ls}

