% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pairs.custom}
\alias{pairs.custom}
\title{Customization of pairs function}
\usage{
\method{pairs}{custom}(df, main = NULL, flag = 0.5)
}
\arguments{
\item{Customization}{of pairs function. Diagonal displays histograms. Lower panel displays pearson's r, with values above 0.5 or other specified value flagged by coloring red.}
}
\description{
Customization of pairs function
}
\examples{
pairs.custom()
}
\keyword{correlation}
\keyword{pairs,}

