% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to.month}
\alias{to.month}
\title{Associates calendar dates with a date representing its month.}
\usage{
to.month(d)
}
\arguments{
\item{Given}{a date, determines which month it falls into, and assigns the date corresponding to the first calendar day of the month.}
}
\description{
Associates calendar dates with a date representing its month.
}
\examples{
to.month()
}
\keyword{dates}

