% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to.season}
\alias{to.season}
\title{Associate calendar dates either a date or string representing its season.}
\usage{
to.season(d, return.type = "date")
}
\arguments{
\item{Given}{a date, determines which season it falls into, and assigns it either a string of the season name,}
}
\description{
Associate calendar dates either a date or string representing its season.
}
\examples{
to.season()
}
\keyword{dates,}
\keyword{season}

