% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to.water.year}
\alias{to.water.year}
\title{Associate calendar dates a date object or numeric year, representing its water year.}
\usage{
to.water.year(d, date.only = T)
}
\arguments{
\item{Given}{a date, determines which water year it falls into, and assigns it either a date corresponding to the last calendar day of the water year, or a numeric representation of the year.}
}
\description{
Associate calendar dates a date object or numeric year, representing its water year.
}
\examples{
to.water.year()
}
\keyword{dates,}
\keyword{water}
\keyword{year}

