% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weather.retrieve}
\alias{weather.retrieve}
\title{import weather}
\usage{
weather.retrieve(gages.spatial, periods = c("seasonal", "annual"),
  agg.function.weather = (conteStreamflow::agg.function.weather),
  template.date = NULL, template.period = NULL, cols.weather = NULL)
}
\description{
import weather
}

